Done: Schedule Submit Menu

+++ resources/views/settings/roles/action.blade.php
@@ -1,6 +1,10 @@
 @if($name === Constant::ADMIN)
   <span class="badge text-primary">{{ trans('Tidak Bisa Diubah') }}</span>
 @else
-  <a href="{{ route('roles.edit', $uuid) }}" class="text-warning me-2"><i class="fa fa-sm fa-pencil"></i></a>
-  <a href="#" onclick="deleteRole(`{{ route('roles.destroy', $uuid) }}`)" class="text-danger me-2"><i class="fa fa-sm fa-trash"></i></a>
+  @can('roles.edit')
+    <a href="{{ route('roles.edit', $uuid) }}" class="text-warning me-2"><i class="fa fa-sm fa-pencil"></i></a>
+  @endcan
+  @can('roles.destroy')
+    <a href="#" onclick="deleteRole(`{{ route('roles.destroy', $uuid) }}`)" class="text-danger me-2"><i class="fa fa-sm fa-trash"></i></a>
+  @endcan
 @endif
